Move CourseList to CourseUtil
