+++ src/test/java/seedu/address/model/course/CourseListTest.java
@@ -4,37 +4,35 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
+
+import seedu.address.model.util.CourseUtil;
+
 public class CourseListTest {
     @Test
     public void containsValidCourse() {
-        CourseList courseList = new CourseList();
-        assertTrue(courseList.contains("CS1101S"));
+        assertTrue(CourseUtil.contains("CS1101S"));
     }
 
     @Test
     public void containsInvalidCourse() {
-        CourseList courseList = new CourseList();
-        assertFalse(courseList.contains("CS9999"));
+        assertFalse(CourseUtil.contains("CS9999"));
     }
 
     @Test
     public void containsNullCourse() {
-        CourseList courseList = new CourseList();
-        assertFalse(courseList.contains(null));
+        assertFalse(CourseUtil.contains(null));
     }
 
     @Test
     public void containsMultipleCourses() {
-        CourseList courseList = new CourseList();
-        assertTrue(courseList.contains("CS1101S"));
-        assertTrue(courseList.contains("MA1521"));
-        assertFalse(courseList.contains("CS9999"));
+        assertTrue(CourseUtil.contains("CS1101S"));
+        assertTrue(CourseUtil.contains("MA1521"));
+        assertFalse(CourseUtil.contains("CS9999"));
     }
 
     @Test
     public void containsCourseWithSpecialCharacters() {
-        CourseList courseList = new CourseList();
-        assertFalse(courseList.contains("CS-1101S"));
+        assertFalse(CourseUtil.contains("CS-1101S"));
     }
 
 }
